+++ src/pages/RFOX/components/Claim/ClaimSelect.tsx
@@ -1,5 +1,5 @@
 import { Button, CardBody, Center, Flex, Skeleton, Stack } from '@chakra-ui/react'
-import { type ChainId, foxAssetId, fromAccountId } from '@shapeshiftoss/caip'
+import { foxAssetId, fromAccountId } from '@shapeshiftoss/caip'
 import dayjs from 'dayjs'
 import { type FC, useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
@@ -7,15 +7,15 @@ import { useHistory } from 'react-router'
 import { AssetIcon } from 'components/AssetIcon'
 import { ClaimStatus } from 'components/ClaimRow/types'
 import { SlideTransition } from 'components/SlideTransition'
-import { RawText, Text } from 'components/Text'
+import { Text } from 'components/Text'
 import { fromBaseUnit } from 'lib/math'
-import { chainIdToChainDisplayName } from 'lib/utils'
 import { useGetUnstakingRequestsQuery } from 'pages/RFOX/hooks/useGetUnstakingRequestsQuery'
 import { useRFOXContext } from 'pages/RFOX/hooks/useRfoxContext'
 import { RfoxTabIndex } from 'pages/RFOX/Widget'
 import { selectAssetById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
+import { ChainNotSupported } from '../Shared/ChainNotSupported'
 import { ClaimRow } from './ClaimRow'
 import type { RfoxClaimQuote } from './types'
 import { ClaimRoutePaths, type ClaimRouteProps } from './types'
@@ -51,28 +51,6 @@ const NoClaimsAvailable: FC<NoClaimsAvailableProps> = ({ isError, setStepIndex }
           {translate('RFOX.unstakeNow')}
         </Button>
       )}
-    </Center>
-  )
-}
-
-type ChainNotSupportedProps = {
-  chainId: ChainId | undefined
-}
-
-const ChainNotSupported: FC<ChainNotSupportedProps> = ({ chainId }) => {
-  const translate = useTranslate()
-
-  if (!chainId) return null
-
-  const chainLabel = chainIdToChainDisplayName(chainId)
-
-  return (
-    <Center flexDir={'column'}>
-      <AssetIcon size='lg' assetId={foxAssetId} showNetworkIcon={false} mb={4} />
-      <Text translation='RFOX.noSupportedChains' fontSize='xl' fontWeight={'bold'} mb={4} />
-      <RawText fontSize='md' color='gray.400' mb={4} textAlign={'center'}>
-        {translate('RFOX.noSupportedChainsDescription', { chainLabel })}
-      </RawText>
     </Center>
   )
 }
